Complete 05-mail-and-img

+++ src/controllers/authController.js
@@ -1,72 +1,47 @@
-import bcrypt from 'bcrypt';
-import createHttpError from 'http-errors';
-
 import { User } from '../models/user.js';
-import { Session } from '../models/session.js';
+import createHttpError from 'http-errors';
+import bcrypt from 'bcrypt';
 import { createSession, setSessionCookies } from '../services/auth.js';
-
-const clearSessionCookies = (res) => {
-  res.clearCookie('sessionId');
-  res.clearCookie('accessToken');
-  res.clearCookie('refreshToken');
-};
+import { Session } from '../models/session.js';
+import jwt from 'jsonwebtoken';
+import { sendEmail } from '../utils/sendMail.js';
+import handlebars from 'handlebars';
+import path from 'node:path';
+import fs from 'node:fs/promises';
 
 export const registerUser = async (req, res) => {
   const { email, password } = req.body;
-
   const existingUser = await User.findOne({ email });
-
   if (existingUser) {
     throw createHttpError(400, 'Email in use');
   }
-
   const hashedPassword = await bcrypt.hash(password, 10);
-
-  const user = await User.create({
+  const newUser = await User.create({
     email,
     password: hashedPassword,
   });
-
-  const session = await createSession(user._id);
-
-  setSessionCookies(res, session);
-
-  res.status(201).json(user);
+  const newSession = await createSession(newUser._id);
+  setSessionCookies(res, newSession);
+  res.status(201).json(newUser);
 };
 
 export const loginUser = async (req, res) => {
   const { email, password } = req.body;
-
   const user = await User.findOne({ email });
-
   if (!user) {
     throw createHttpError(401, 'Invalid credentials');
   }
-
-  const isPasswordCorrect = await bcrypt.compare(password, user.password);
-
-  if (!isPasswordCorrect) {
+  const isValidPassword = await bcrypt.compare(password, user.password);
+  if (!isValidPassword) {
     throw createHttpError(401, 'Invalid credentials');
   }
-
-  await Session.deleteMany({
-    userId: user._id,
-  });
-
-  const session = await createSession(user._id);
-
-  setSessionCookies(res, session);
-
+  await Session.deleteOne({ userId: user._id });
+  const newSession = await createSession(user._id);
+  setSessionCookies(res, newSession);
   res.status(200).json(user);
 };
-
 export const refreshUserSession = async (req, res) => {
   const { sessionId, refreshToken } = req.cookies;
-
-  if (!sessionId || !refreshToken) {
-    clearSessionCookies(res);
-    throw createHttpError(401, 'Session not found');
-  }
 
   const session = await Session.findOne({
     _id: sessionId,
@@ -74,25 +49,16 @@ export const refreshUserSession = async (req, res) => {
   });
 
   if (!session) {
-    clearSessionCookies(res);
     throw createHttpError(401, 'Session not found');
   }
 
-  const isRefreshTokenExpired =
-    new Date() > new Date(session.refreshTokenValidUntil);
-
-  if (isRefreshTokenExpired) {
-    await Session.deleteOne({
-      _id: session._id,
-    });
-
-    clearSessionCookies(res);
-
-    throw createHttpError(401, 'Refresh token expired');
+  const isTokenExpired = new Date() > new Date(session.refreshTokenValidUntil);
+  if (isTokenExpired) {
+    throw createHttpError(401, 'Session token expired');
   }
 
   await Session.deleteOne({
-    _id: session._id,
+    _id: sessionId,
   });
 
   const newSession = await createSession(session.userId);
@@ -100,20 +66,84 @@ export const refreshUserSession = async (req, res) => {
   setSessionCookies(res, newSession);
 
   res.status(200).json({
-    message: 'Successfully refreshed a session',
+    message: 'Session refreshed',
   });
 };
 
 export const logoutUser = async (req, res) => {
   const { sessionId } = req.cookies;
-
   if (sessionId) {
     await Session.deleteOne({
       _id: sessionId,
     });
   }
+  res.clearCookie('accessToken');
 
-  clearSessionCookies(res);
+  res.clearCookie('refreshToken');
+
+  res.clearCookie('sessionId');
 
   res.status(204).send();
+};
+
+export const requestResetEmail = async (req, res) => {
+  const { email } = req.body;
+  const user = await User.findOne({ email });
+  if (!user) {
+    res.status(200).json({ message: 'Password reset email sent successfully' });
+    return;
+  }
+  const resetToken = jwt.sign(
+    {
+      sub: user._id,
+      email,
+    },
+    process.env.JWT_SECRET,
+    { expiresIn: '15m' },
+  );
+  const templatePath = path.resolve('src/templates/reset-password-email.html');
+  const templateSource = await fs.readFile(templatePath, 'utf-8');
+  const template = handlebars.compile(templateSource);
+  const html = template({
+    name: user.username,
+    link: `${process.env.FRONTEND_DOMAIN}/reset-password?token=${resetToken}`,
+  });
+  try {
+    await sendEmail({
+      from: process.env.SMTP_FROM,
+      to: email,
+      html,
+    });
+  } catch {
+    throw createHttpError(
+      500,
+      'Failed to send the email, please try again later.',
+    );
+  }
+  res.status(200).json({ message: 'Password reset email sent successfully' });
+};
+
+export const resetPassword = async (req, res) => {
+  const { token, password } = req.body;
+  let payload;
+  try {
+    payload = jwt.verify(token, process.env.JWT_SECRET);
+  } catch {
+    throw createHttpError(401, 'Invalid or expired token');
+  }
+  const user = await User.findOne({ _id: payload.sub, email: payload.email });
+  if (!user) {
+    throw createHttpError(404, 'User not found');
+  }
+  const hashedPassword = await bcrypt.hash(password, 10);
+  await User.updateOne(
+    {
+      _id: user._id,
+    },
+    { password: hashedPassword },
+  );
+  await Session.deleteMany({ userId: user._id });
+  res.status(200).json({
+    message: 'Password reset successfully',
+  });
 };
